requirements.txt
Django
sentence-transformers
scikit-learn
numpy
pdfplumber
python-docx
